refactor(e2e): import schema types directly, remove re-export barrel from e2e-reporter
e2e-reporter.ts was re-exporting schema types to let consumers mix
reporter functions and schema types in a single import. Remove the
barrel; consumers now import schema types from schemas/e2e-report.schema
and reporter-own items from e2e-reporter directly.
